Update title in the readme
"GLib/GTK+ Application Development" would have the same acronym as GGAD
written by Havoc Pennington, so I'm not sure if it would be a good idea.

The new title doesn't imply that it's only for applications. There could
be a chapter on how to write a library, or an operating system service.