Better listings and figures placement
For listings, I prefer the caption at the bottom, because most Listings
have at the top an #include, so when we read and see an #include, it's
clear that the #include is part of the listing. Moreover when there is a
listing followed directly by a figure, it's clearer when both captions
are at the bottom.

For the float placement, I think it's better like that, to let LaTeX do
its job as it likes.